refactor: background.js 함수들을 backgroundUtils.js로 모듈화 분리
